+++ core/src/main/java/com/graphhopper/routing/util/DefaultVehicleEncodedValuesFactory.java
@@ -34,13 +34,13 @@ public class DefaultVehicleEncodedValuesFactory implements VehicleEncodedValuesF
             return VehicleEncodedValues.car(configuration);
 
         if (name.equals("car4wd"))
-            throw new IllegalArgumentException("Instead of car4wd use the roads vehicle and a custom_model, see custom_models/car4wd.json");
+            throw new IllegalArgumentException("Instead of car4wd use custom_models/car4wd.json");
 
         if (name.equals(BIKE))
             return VehicleEncodedValues.bike(configuration);
 
         if (name.equals("bike2"))
-            throw new IllegalArgumentException("Instead of bike2 use the bike vehicle and a custom model, see custom_models/bike.json and #1234");
+            throw new IllegalArgumentException("Instead of bike2 use custom_models/bike.json, see #2668");
 
         if (name.equals(RACINGBIKE))
             return VehicleEncodedValues.racingbike(configuration);
@@ -51,11 +51,11 @@ public class DefaultVehicleEncodedValuesFactory implements VehicleEncodedValuesF
         if (name.equals(FOOT))
             return VehicleEncodedValues.foot(configuration);
 
-        if (name.equals(HIKE))
-            throw new IllegalArgumentException("Instead of hike use the foot vehicle and a custom model, see custom_models/hike.json and #2759");
+        if (name.equals("hike"))
+            throw new IllegalArgumentException("Instead of hike use custom_models/hike.json, see #2759");
 
-        if (name.equals(MOTORCYCLE))
-            return VehicleEncodedValues.motorcycle(configuration);
+        if (name.equals("motorcycle"))
+            throw new IllegalArgumentException("Instead of motorcycle use custom_models/motorcycle.json, see #2781");
 
         if (name.equals(WHEELCHAIR))
             return VehicleEncodedValues.wheelchair(configuration);