+++ core/src/main/java/com/graphhopper/routing/util/parsers/RoadsAccessParser.java
@@ -2,12 +2,16 @@ package com.graphhopper.routing.util.parsers;
 
 import com.graphhopper.reader.ReaderWay;
 import com.graphhopper.routing.ev.BooleanEncodedValue;
-import com.graphhopper.routing.ev.EncodedValueLookup;
 import com.graphhopper.routing.ev.EdgeIntAccess;
+import com.graphhopper.routing.ev.EncodedValueLookup;
 import com.graphhopper.routing.ev.VehicleAccess;
 import com.graphhopper.storage.IntsRef;
 import com.graphhopper.util.PMap;
 
+/**
+ * Access parser (boolean) for the 'roads' vehicle. Not to be confused with OSMRoadAccessParser that fills road_access
+ * enum (for car).
+ */
 public class RoadsAccessParser implements TagParser {
     private final BooleanEncodedValue accessEnc;