+++ core/src/test/java/com/graphhopper/routing/util/parsers/AbstractBikeTagParserTester.java
@@ -540,4 +540,132 @@ public abstract class AbstractBikeTagParserTester {
         assertTrue(bike.intendedValues.contains("private"));
         assertFalse(bike.isBarrier(node));
     }
+
+    @Test
+    public void testOneway() {
+        ReaderWay way = new ReaderWay(1);
+        way.setTag("highway", "tertiary");
+        assertAccess(way, true, true);
+
+        way.setTag("oneway", "yes");
+        assertAccess(way, true, false);
+
+        way.clearTags();
+        way.setTag("highway", "tertiary");
+        way.setTag("oneway:bicycle", "yes");
+        assertAccess(way, true, false);
+
+        way.clearTags();
+        way.setTag("highway", "tertiary");
+        way.setTag("oneway", "yes");
+        way.setTag("oneway:bicycle", "no");
+        assertAccess(way, true, true);
+
+        way.clearTags();
+        way.setTag("highway", "tertiary");
+        way.setTag("oneway", "yes");
+        way.setTag("oneway:bicycle", "-1");
+        assertAccess(way, false, true);
+
+        way.clearTags();
+        way.setTag("highway", "tertiary");
+        way.setTag("oneway", "yes");
+        way.setTag("cycleway:right:oneway", "no");
+        assertAccess(way, true, true);
+
+        way.clearTags();
+        way.setTag("highway", "tertiary");
+        way.setTag("oneway", "yes");
+        way.setTag("cycleway:right:oneway", "-1");
+        assertAccess(way, false, true);
+
+        way.clearTags();
+        way.setTag("highway", "tertiary");
+        assertAccess(way, true, true);
+
+        way.clearTags();
+        way.setTag("highway", "tertiary");
+        way.setTag("vehicle:forward", "no");
+        assertAccess(way, false, true);
+
+        way.clearTags();
+        way.setTag("highway", "tertiary");
+        way.setTag("bicycle:forward", "no");
+        assertAccess(way, false, true);
+
+        way.clearTags();
+        way.setTag("highway", "tertiary");
+        way.setTag("vehicle:backward", "no");
+        assertAccess(way, true, false);
+
+        way.clearTags();
+        way.setTag("highway", "tertiary");
+        way.setTag("motor_vehicle:backward", "no");
+        assertAccess(way, true, true);
+
+        way.clearTags();
+        way.setTag("highway", "tertiary");
+        way.setTag("oneway", "yes");
+        way.setTag("bicycle:backward", "no");
+        assertAccess(way, true, false);
+
+        way.setTag("bicycle:backward", "yes");
+        assertAccess(way, true, true);
+
+        way.clearTags();
+        way.setTag("highway", "residential");
+        way.setTag("oneway", "yes");
+        way.setTag("bicycle:backward", "yes");
+        assertAccess(way, true, true);
+
+        way.clearTags();
+        way.setTag("highway", "residential");
+        way.setTag("oneway", "-1");
+        way.setTag("bicycle:forward", "yes");
+        assertAccess(way, true, true);
+
+        way.clearTags();
+        way.setTag("highway", "tertiary");
+        way.setTag("bicycle:forward", "use_sidepath");
+        assertAccess(way, true, true);
+
+        way.clearTags();
+        way.setTag("highway", "tertiary");
+        way.setTag("bicycle:forward", "use_sidepath");
+        assertAccess(way, true, true);
+
+        way.clearTags();
+        way.setTag("highway", "tertiary");
+        way.setTag("oneway", "yes");
+        way.setTag("cycleway", "opposite");
+        assertAccess(way, true, true);
+
+        way.clearTags();
+        way.setTag("highway", "residential");
+        way.setTag("oneway", "yes");
+        way.setTag("cycleway:left", "opposite_lane");
+        assertAccess(way, true, true);
+
+        way.clearTags();
+        way.setTag("highway", "secondary");
+        way.setTag("cycleway:left:oneway", "-1");
+        way.setTag("cycleway:both", "track");
+        assertAccess(way, true, true);
+
+        way.clearTags();
+        way.setTag("highway", "secondary");
+        way.setTag("oneway", "yes");
+        way.setTag("cycleway:both", "no");
+        assertAccess(way, true, false);
+    }
+
+    private void assertAccess(ReaderWay way, boolean fwd, boolean bwd) {
+        EdgeIntAccess edgeIntAccess = new ArrayEdgeIntAccess(1);
+        int edge = 0;
+        IntsRef edgeFlags = new IntsRef(1);
+        IntsRef relationFlags = new IntsRef(1);
+        accessParser.handleWayTags(edge, edgeIntAccess, way, relationFlags);
+        if (fwd) assertTrue(accessEnc.getBool(false, edge, edgeIntAccess));
+        if (bwd) assertTrue(accessEnc.getBool(true, edge, edgeIntAccess));
+    }
 }